More refactoring DirectSBS, commonVR
