+++ impl11/ddraw/DirectSBS.h
@@ -1,0 +1,4 @@
+#pragma once
+
+
+bool InitDirectSBS();
